refactor: remove landing component barrel
