--- apps/landing/src/components/landing/index.ts
@@ -1,6 +1,0 @@
-export { LandingHeader } from "./header";
-export { LandingFooter } from "./footer";
-export { HeroSection } from "./hero-section";
-export { FaqPanel } from "./faq-panel";
-export { Testimonials } from "./testimonials";
-export { HomeContent } from "./home-content";
